animator: use double for time accumulator

+++ src/components/animator.hpp
@@ -7,7 +7,7 @@
 #include "renderer/animation.hpp"
 
 struct Animator {
-    float                      Time;
+    double                     Time;
     std::string                AnimationName;
     std::optional<::Animation> Animation;
     std::vector<glm::mat4>     FinalBonesMatrices;
